zeromq/pyproject.toml
[tool.poetry]
name = "streaming"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyzmq = "^19.0.1"
click = "^7.1.2"
click-pathlib = "^2020.3.13"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
streaming = "streaming:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

